[CL-2084] Move folder moderator notification outlet to main app
