+++ front/app/containers/MainHeader/NotificationMenu/components/ProjectFolderModerationRightsReceivedNotification/index.tsx
