+++ front/app/containers/MainHeader/NotificationMenu/components/ProjectFolderModerationRightsReceivedNotification/messages.ts
